Cleaning some
